Replace remote URL with ssh instead of https
Using the https URL in the manifest requires capturing credentials from the command line or in .netrc to work correctly.
The former doesn't play nicely with automatic builds and the latter has to be kept up to date as your password changes. (meh!)
If you prefer using ssh, you have to use a .gitconfig insteadOf to re-write the URL to use ssh for each account.

Signed-off-by: Sean Hudson <sean_hudson@mentor.com>