amd: Use master branch
Our release/2014.12 branch is not yet usable.  Neither for
that matter is our master branch but we are currently doing
our work there.  To avoid having to work in two branches at
once we are currently trying to get master working.  Once it
builds sanely then we will create release/2014.12 and update
the manifest.

Signed-off-by: Drew Moseley <drew_moseley@mentor.com>