amd_dev.xml: fetch mel-updates
The AMD BSPs are not in sync with the standard releases and
there are changes that are targetted to go into the standard
mel repos. In order to align with the process these changes
will be pushed through mel-updates so it can be picked up
by the product team as needed.

Signed-off-by: Awais Belal <awais_belal@mentor.com>